+++ OOP Exam Quiz/Player.h
@@ -2,7 +2,7 @@
 class Player : public User
 {
 	string _latestPlayedQuiz = "";
-	float _highestScore = 0;
+	double _highestScore = 0;
 
 public:
